requirements.txt
Flask==3.0.2
mysql-connector-python==9.0.0
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2
Flask-Cors==4.0.0
passlib==1.7.4
firebase-admin==6.5.0
Werkzeug==3.0.2
python-dateutil==2.9.0.post0
numpy==1.26.4
redis==5.1.0
